Significant changes made
fill color changed for genre count bar charts.
legends removed for genre counts bar chart.
genre count tables added for tv shows and movies.
new table visualization created for combined Netflix and Disney Plus datasets. The repeated titles are counted shows whether it is a tv show or movie, and shows which countries they are available. Point of the table is to show the overlap of content between Netflix and Disney Plus.